Final version for FinalProject.ino
Reworked from version 6-7 and adjusted any logic implementation as needed to adhere to rubric guidelines. For the final version I added the stepper motor for the vent and fine tuned the code as needed to hammer out any bugs. This version encompasses all of the necessary requirement's and restrictions.
It functions as desired for the swamp cooler.